Game loop priority.
Game loops can now be sorted by priority, which ist just a signed 32 bit integer.
I refrained from implemented a full-blown dependency management system as it is probably overkill for us.
There is a convenience function to print the tick order, which can be used for debugging (kr::GlobalGameLoopRegistry::printTickOrder).